Handle `if n:` for an optional argument `n` (#707)
As mentioned in the discussion of issue #700, optional arguments are only tested for existence when we write `n is None` explicitly, however it should also be tested if the argument is cast to a bool (as in the expression `if n:`). This PR fixes this

**Commit Summary**

- Split pyccel.ast.operators into 2 files to avoid circular imports between pyccel.ast.operators and  pyccel.ast.builtins
    - The new file pyccel.ast.bitwise_operators is created
    - It requires the `PythonInt` cast
- Add missing classes to pyccel.ast.operators.__all__
- Handle boolean cast of optional argument in `PythonBool.__new__`
    - When the argument of `PythonBool` is an optional object the `__new__` function modifies the result to check that the argument exists:
    `bool(a)`->`a is not None and bool(a)`
    - Handling this in the `__new__` function ensures that operator precedence is handled with minimal parentheses even when the `PythonBool` is created in the constructor of `pyccel.ast.core.If`
- Add a test.
