+++ pyccel/ast/builtins.py
@@ -20,8 +20,10 @@ from .datatypes import (NativeInteger, NativeBool, NativeReal,
                         NativeComplex, NativeString, str_dtype,
                         NativeGeneric, default_precision)
 from .internals import PyccelInternalFunction
-from .literals  import LiteralInteger, LiteralFloat, LiteralComplex
+from .literals  import LiteralInteger, LiteralFloat, LiteralComplex, Nil
 from .literals  import Literal, LiteralImaginaryUnit, get_default_literal_value
+from .operators import PyccelAdd, PyccelAnd, PyccelMul, PyccelIsNot
+from .operators import PyccelMinus, PyccelUnarySub
 
 __all__ = (
     'PythonReal',
@@ -125,7 +127,10 @@ class PythonBool(Expr, PyccelAstNode):
     _dtype = NativeBool()
 
     def __new__(cls, arg):
-        return Expr.__new__(cls, arg)
+        if getattr(arg, 'is_optional', None):
+            return PyccelAnd(PyccelIsNot(arg, Nil()), Expr.__new__(cls, arg))
+        else:
+            return Expr.__new__(cls, arg)
 
     @property
     def arg(self):
@@ -172,7 +177,6 @@ class PythonComplex(Expr, PyccelAstNode):
 
         # Split arguments depending on their type to ensure that the arguments are
         # either a complex and LiteralFloat(0) or 2 floats
-        from .operators import PyccelAdd, PyccelMul
 
         if arg0.dtype is NativeComplex() and arg1.dtype is NativeComplex():
             # both args are complex
@@ -189,7 +193,6 @@ class PythonComplex(Expr, PyccelAstNode):
             self._internal_var = arg0
 
         else:
-            from .operators import PyccelAdd, PyccelMinus, PyccelUnarySub
 
             if arg0.dtype is NativeComplex() and \
                     not (isinstance(arg1, Literal) and arg1.python_value == 0):
